+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 # intellij-direnv Changelog
 
 ## [Unreleased]
+## [0.2.0]
 ### Added
 - Allow importing non-toplevel `.envrc` files